feat:Change PDF template editor to show a full screen editor instead of a modal
