+++ apps/tenant-management-webapp/src/app/app.tsx
@@ -17,7 +17,7 @@ import { PrivateApp, PrivateRoute } from './privateApp';
 import { fetchConfig } from '@store/config/actions';
 import AuthContext from '@lib/authContext';
 import CreateTenant from '@pages/admin/tenants/CreateTenant';
-
+import { EditorLayoutManager } from '@pages/editor';
 import { ThemeProvider } from 'styled-components';
 import { theme } from 'theme';
 import PublicApp from './publicApp';
@@ -46,7 +46,11 @@ const AppRouters = () => {
                 <PrivateRoute path="/admin/tenants" component={TenantsRouter} />
               </PrivateApp>
             </Route>
-
+            <Route path="/editor">
+              <PrivateApp>
+                <PrivateRoute path="/editor" component={EditorLayoutManager} />
+              </PrivateApp>
+            </Route>
             <PublicApp>
               <Route path="/:realm/login">
                 <Login />
