+++ apps/tenant-management-webapp/src/app/pages/admin/services/pdf/templates/utils/suggestion.ts
@@ -1,0 +1,24 @@
+export const getSuggestion = () => [
+  {
+    label: 'data',
+    insertText: 'data',
+    children: [
+      {
+        label: 'id',
+        insertText: 'id',
+      },
+      {
+        label: 'name',
+        insertText: 'name',
+      },
+      {
+        label: 'description',
+        insertText: 'description',
+      },
+      {
+        label: 'template',
+        insertText: 'template',
+      },
+    ],
+  },
+];
